Refactor Diff cmd application logic to pkg/, some cleanup
