+++ pkg/kubecfg/delete.go
@@ -1,3 +1,18 @@
+// Copyright 2017 The kubecfg authors
+//
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
 package kubecfg
 
 import (
@@ -7,30 +22,39 @@ import (
 	log "github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 
+	"github.com/ksonnet/kubecfg/metadata"
+	"github.com/ksonnet/kubecfg/template"
 	"github.com/ksonnet/kubecfg/utils"
 )
 
+// DeleteCmd represents the delete subcommand
 type DeleteCmd struct {
 	ClientPool       dynamic.ClientPool
 	Discovery        discovery.DiscoveryInterface
 	DefaultNamespace string
 
-	GracePeriod int64
+	Expander    *template.Expander
+	Environment *string
+	Files       []string
 
-	Objs []*unstructured.Unstructured
+	GracePeriod int64
 }
 
-func (c DeleteCmd) Run() error {
+func (c DeleteCmd) Run(wd metadata.AbsPath) error {
 	version, err := utils.FetchVersion(c.Discovery)
 	if err != nil {
 		return err
 	}
 
-	sort.Sort(sort.Reverse(utils.DependencyOrder(c.Objs)))
+	objs, err := c.Expander.Expand(c.Files)
+	if err != nil {
+		return err
+	}
+
+	sort.Sort(sort.Reverse(utils.DependencyOrder(objs)))
 
 	deleteOpts := metav1.DeleteOptions{}
 	if version.Compare(1, 6) < 0 {
@@ -46,7 +70,7 @@ func (c DeleteCmd) Run() error {
 		deleteOpts.GracePeriodSeconds = &c.GracePeriod
 	}
 
-	for _, obj := range c.Objs {
+	for _, obj := range objs {
 		desc := fmt.Sprintf("%s %s", utils.ResourceNameFor(c.Discovery, obj), utils.FqName(obj))
 		log.Info("Deleting ", desc)
